Rename util/misc to util/ts
